Update Dockerfile and tweak main.cpp
This changes some of the dockerfile to include different build steps including an option commented out way of getting the code into the container (more portable if you're running this in a host that doesnt have the repo cloned - it does make iteration different on your local machine). This also adjusts main.cpp to reference the installed library for libds at the right reference point.